Added:FindingMaxOfFloat

+++ src/main/java/org/blz/Generics/Maximum.java
@@ -1,14 +1,9 @@
 package org.blz.Generics;
 
 public class Maximum {
-//    public static void main(String[] args) {
-//        Integer a=11,b=7,c=9;
-//        getMaximum(a,b,c);
-//
-//        }
 
-    public static Integer getMaximum(Integer a, Integer b, Integer c) {
-        Integer max=a;
+    public static Float getMaximum(Float a, Float b, Float c) {
+        Float max=a;
             if(b.compareTo(a)>0 && b.compareTo(c)>0)
             {
             max=b;
